tests/wpt/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.translate.3d.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.transformation.translate.3d</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.translate.3d</h1>
<p class="desc">translate() function with three arguments modifies the underlying matrix appropriately</p>


<script>
var t = async_test("translate() function with three arguments modifies the underlying matrix appropriately");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

const dx = 2;
const dy = 3;
const dz = 4;
ctx.translate(dx, dy, dz);
let canvasTransform = ctx.getTransform();
_assert(canvasTransform.m41 = dx, "canvasTransform.m41 = dx");
_assert(canvasTransform.m42 = dy, "canvasTransform.m42 = dy");
_assert(canvasTransform.m43 = dz, "canvasTransform.m43 = dz");
ctx.translate(dx, dy, dz);
canvasTransform = ctx.getTransform();
_assert(canvasTransform.m41 = 2 * dx, "canvasTransform.m41 = 2 * dx");
_assert(canvasTransform.m42 = 2 * dy, "canvasTransform.m42 = 2 * dy");
_assert(canvasTransform.m43 = 2 * dz, "canvasTransform.m43 = 2 * dz");
t.done();

});
</script>
